websocket-server/requirements.txt
python-socketio
gevent